feat(android): create AudioUtils
